[Feature #163083232] Refactor API controllers to use data from db
