+++ server/routes/meetup.js
@@ -25,8 +25,8 @@ router.get('/meetups/:meetupId/rsvps', meetupController.getAllRsvps);
 
 router
   .route('/meetups/:meetupId/questions/:questionId')
-  .get(meetupController.getSingleMeetupQuestion)
-  .patch(meetupController.updateMeetupQuestion)
-  .delete(meetupController.deleteMeetupQuestion);
+  .get(questionController.getSingleMeetupQuestion)
+  .patch(questionController.updateMeetupQuestion)
+  .delete(questionController.deleteMeetupQuestion);
 
 export default router;
